fix: Revert image redirect #173
