--- src/Fan.Web/Pages/Image.cshtml
@@ -1,5 +1,0 @@
-﻿@page "{appType}/{size}/{userId:int}/{year}/{month}/{fileName}"
-@model Fan.Web.Pages.ImageModel
-@{
-    Layout = null;
-}
